+++ src/comments/getFilePathFromError.ts
@@ -1,0 +1,1 @@
+export const getFilePathFromError = (error: string) => error.split('(')[0];