+++ CHANGELOG.md
@@ -10,7 +10,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 - Strict by default without `@ts-strict` comment
 - Ignores file with `@ts-strict-ignore` comment
-- new flag in cli `--updateComments` which updates comments based on strict errors
+- new flag in argv `--updateComments` which updates comments based on strict errors
 
 ## [1.1.2] - 2021-06-15
 
@@ -28,7 +28,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- Adds cli tool to check strict files during build time
+- Adds argv tool to check strict files during build time
 
 ## [1.0.1] - 2021-06-01